requirements.txt
google-adk
praw
python-dotenv 
litellm
uvx<2.0
